correzione di alcuni metodi
